added ExCC as 2nd method & slightly modified the code
ExCC is a method exact for solving CC problem.
Currently, ExCC is slow (it takes 5-6 hours for a graph with 600 nodes).
Consequently, it is preferable to apply it on graphs up to 100-200 nodes.
TODO: Cplex is used with Open-Jdk. So the jar file called "cplex-partition-openjdk-1_8.jar"
is not compatible with oracle Jdk.