update plot.graph.stats according to network.no.list && change 'out' folder structure
Since we generate many graphs with the same parameter set, we updated plot.grap.stats().
Now, it generates the same plots as before, but for all networks (according to network.no.list)
On the other hand, we changed 'out' folder structure: out/networks and out/plots.
Lastly, we found a bug in plot.graph.stats(): legend was buggy when x-axis represents 'prop.negs', we corrected it